rename OpenAuthoring > OpenAuthoringProvider
