+++ open-authoring/useUpdateAuthOnFormFail.ts
@@ -1,7 +1,7 @@
 import { Form } from 'tinacms'
 import { useEffect } from 'react'
 import createDecorator from 'final-form-submit-listener'
-import { useOpenAuthoring } from '../components/layout/OpenAuthoring'
+import { useOpenAuthoring } from './OpenAuthoringProvider'
 
 // Hook to update root openAuthoring state when form fails.
 // We need to perform to check before an action is clicked (e.g start auth flow)
